+++ src/parser/Ast.ts
@@ -386,14 +386,16 @@ export interface ProcedureAst extends AstWithInternalWhitespace {
   type: AstNodeType.FUNCTION_CALL,
   procedureName: string,
   args: Ast[],
+  hyperlink?: string,
 }
 
-export const buildProcedureAst = (procedureName: string, args: Ast[], leadingWhitespace?: IToken, internalWhitespace?: IToken): ProcedureAst => ({
+export const buildProcedureAst = (procedureName: string, args: Ast[], leadingWhitespace?: IToken, internalWhitespace?: IToken, hyperlink?: string): ProcedureAst => ({
   type: AstNodeType.FUNCTION_CALL,
   procedureName,
   args,
   leadingWhitespace: leadingWhitespace?.image,
   internalWhitespace: internalWhitespace?.image,
+  hyperlink
 })
 
 export interface ArrayAst extends AstWithInternalWhitespace {