+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added HYPERLINK function.
+  [#1215](https://github.com/handsontable/hyperformula/issues/1215)
+
+### Fixed
+
+- Fixed the `SEARCH` function to be case-insensitive regardless of the engine configuration.
+  [#1225](https://github.com/handsontable/hyperformula/issues/1225)
+
 ## [2.4.0] - 2023-04-24
 
 ### Added